fix: 퀴즈 답안별 카운트 모델 제공 쿼리 수정

+++ api/src/main/java/com/tdns/toks/api/domain/quiz/model/QuizReplyCountsModel.java
@@ -8,13 +8,13 @@ import java.util.Map;
 @Data
 @AllArgsConstructor
 public class QuizReplyCountsModel {
-    private Integer totalCount;
+    private Long totalCount;
     private Map<String, ReplyModel> replyCount;
 
     @Data
     @AllArgsConstructor
     public static class ReplyModel {
         private String answer;
-        private Integer count;
+        private Long count;
     }
 }
