+++ core/src/main/java/com/tdns/toks/core/domain/quiz/model/QuizReplyCountModel.java
@@ -3,5 +3,5 @@ package com.tdns.toks.core.domain.quiz.model;
 public interface QuizReplyCountModel {
     String getAnswer();
 
-    Integer getCount();
+    Long getCount();
 }
